Did more investigations and it turned out that my previous attempt at an additional regulatisation term if the correct neuron does not spike at all didn't work because of me misunderstanding the timing situation (i.e. which is the last timestep in a trial). I now implemented the extra cost term by triggering "phantom backspikes".
Also changed the type of regularisation term from exponential to 1/x. This way the regularization only kicks in
for very late spikes and for the phantom spikes.

The parameters are only roughly tuned but now achieve some 97% correct. And very low loss (even though not 100%
comparable due to the changed regularization term).

Through experimentation I have convinced myself that the remaining issues are simply that there is a fine balance
between strength of regularisation and the other loss terms. The failures remaining are essentially the non-spike
responses (no output neuron spikes). However, with the regularisation described in the paper increased, there is too
much distortion that leads to other errors.

Even with my phantom spikes, there is still a non-zero chance that the learning becomes unstable and neurons
fall silent. Some parameter tuning might help with that.

Last thing - to get beyond the 94% hurdle it was also important to extend the period of the Adam optimizer a bit
(0.9 -> 0.98, 0.999 -> 0.9998) ... this way learning is not clamped down too quickly; the parameter choices described
in the paper suggest based on my experimentation that the authors have optimised the adam parameters and learning
rate decay so that they just about steer clear from the cliff of silent neurons. I personally think that isn't
the most useful way of setting things up - ideally one would want to be able to let learning run its course without
having to worry about some mal-adapted absorbing fixed point (to be avoided by curtailing the travel in weight
space through Adam parameters and learnign rate decay).